# apps/web/app/game-play/_hooks/types.ts
import type { components } from "@/lib/api-generated/api";

export type GameDto = components["schemas"]["GameDto"];
export type FrameDto = components["schemas"]["FrameDto"];
export type CreateGameDto = components["schemas"]["CreateGameDto"];